Refactoring: put Flask app in a separate subfolder and invoke it via run.py
