+++ backend/app/app.py
@@ -1,0 +1,35 @@
+from flask import Flask, request, jsonify
+from database import connect_to_mongodb
+from db_utils import *
+from bson import json_util
+import json
+
+def create_app():
+    app = Flask(__name__)
+    client, db = connect_to_mongodb()
+
+    @app.route('/movies', methods=['GET'])
+    def get_movies():
+        movies = get_all_movies(db)
+        return json.loads(json_util.dumps(movies))
+
+    @app.route('/movies/<int:tmdb_id>', methods=['GET'])
+    def get_movie(tmdb_id: int):
+        movie = get_movie_by_id(db, tmdb_id)
+        if movie:
+            return json.loads(json_util.dumps(movie))
+        return jsonify({"error": "Movie not found"}), 404
+
+    @app.route('/movies', methods=['POST'])
+    def add_movie():
+        result = insert_movie(db, request.json)
+        return jsonify(result), 201
+
+    @app.route('/movies/<int:tmdb_id>', methods=['DELETE'])
+    def delete_movie(tmdb_id: int):
+        result = delete_movie_by_id(db, tmdb_id)
+        if result['deleted_count'] > 0:
+            return jsonify({"message": "Movie deleted successfully"}), 200
+        return jsonify({"error": "Movie not found"}), 404
+
+    return app
